separated real num traits & cast from real num header
