+++ include/SimpleObjects/RealNumCompare.hpp
